Move parsing of groups and drafts a utility
