+++ app/routes/print-overviews/decisions/overview.js
@@ -2,14 +2,15 @@ import Route from '@ember/routing/route';
 import { inject } from '@ember/service';
 import { hash } from 'rsvp';
 import EmberObject from '@ember/object';
-import {
-  setCalculatedGroupPriorities,
-  groupAgendaitemsByGroupname,
-} from 'fe-redpencil/utils/agenda-item-utils';
+import { parseDraftsAndGroupsFromAgendaitems } from 'fe-redpencil/utils/agenda-item-utils';
 
 export default Route.extend({
   agendaService: inject(),
   type: 'decisions',
+
+  queryParams: {
+    definite: { refreshModel: false }
+  },
 
   async model() {
     const session = await this.modelFor('print-overviews');
@@ -24,7 +25,7 @@ export default Route.extend({
       return item.showAsRemark;
     });
 
-    const { draftAgendaitems, groupedAgendaitems } = await this.parseAgendaItems(agendaitems);
+    const { draftAgendaitems, groupedAgendaitems } = await parseDraftsAndGroupsFromAgendaitems(agendaitems);
 
     await this.agendaService.groupAgendaItemsOnGroupName(draftAgendaitems);
 
@@ -52,20 +53,4 @@ export default Route.extend({
       meeting: session,
     });
   },
-
-  async parseAgendaItems(agendaitems) {
-    const draftAgendaitems = agendaitems.filter((item) => !item.showAsRemark && !item.isApproval);
-
-    await setCalculatedGroupPriorities(draftAgendaitems);
-
-    const groupedAgendaitems = Object.values(groupAgendaitemsByGroupname(draftAgendaitems));
-    return {
-      draftAgendaitems,
-      groupedAgendaitems,
-    };
-  },
-
-  queryParams: {
-    definite: { refreshModel: false }
-  }
 });
